chatbot/requirements.txt
fastapi
uvicorn
torch
transformers
jinja2
pyyaml
